requirements.txt
configparser
Django
django-debug-toolbar
django-extra-views
django-test-plus
flup
Pillow
reportlab
six
sqlparse
